Implemented time taken

+++ Application/AiTemplates/Dtos/EnglishToSqlDto.cs
@@ -3,6 +3,6 @@ namespace Application.AiTemplates.Dtos;
 public class EnglishToSqlDto
 {
     public required string Response { get; set; }
-    public float TimeTaken { get; set; }
+    public double TimeTaken { get; set; }
     public int TokenCost { get; set; }
 }
